Moved Header under common components
